// src/types/device.ts
interface Device {
    id: number;
    name: string;
    device_type: DeviceType;
}

enum DeviceType {
    Unknown = 0,
    Switch = 1,
    Dimmer = 2,
    Actuator = 3,
    MotionSensor = 4,
    ContactSensor = 5,
    DoorbellAcdb7000a = 6,
    WallControl1Channel = 7,
    WallControl2Channel = 8,
    RemoteControl1Channel = 9,
    RemoteControl2Channel = 10,
    RemoteControl3Channel = 11,
    RemoteControl16Channel = 12,
    RemoteControlAyct202 = 15,
    RemoteControlAyct303 = 55,
    Chime = 13,
    DuskSensor = 14,
    ArcRemote = 16,
    ArcContactSensor = 17,
    ArcMotionSensor = 18,
    ArcSmokeSensor = 19,
    ArcSiren = 20,
    DoorbellAcdb7000b = 21,
    Awmt = 22,
    SomfyActuator = 23,
    Light = 24,
    WallSwitchAgst8800 = 25,
    WallSwitchAgst8802 = 26,
    BrelActuator = 27,
    ContactSensor2 = 28,
    ArcKeychainRemote = 29,
    ArcActionButton = 30,
    ArcRotaryDimmer = 31,
    ZigbeeUnknownDevice = 32,
    ZigbeeSwitch = 33,
    ZigbeeDimmer = 34,
    ZigbeeRgb = 35,
    ZigbeeTunable = 36,
    ZigbeeMultiPurposeSensor = 37,
    ZigbeeLock = 38,
    ZigbeeLightLinkRemote = 39,
    ZigbeeLight = 40,
    ZigbeeSocket = 41,
    ZigbeeBuiltInSwitch = 53,
    ZigbeeBuiltInDimmer = 54,
    ZigbeeLeakageSensor = 42,
    ZigbeeSmokeSensor = 43,
    ZigbeeCarbonMonoxideSensor = 47,
    ZigbeeTemperatureAndHumiditySensor = 46,
    ZigbeeLightGroup = 48,
    ZigbeeFireangelSensor = 49,
    DoorbellAcdb8000a = 50,
    ForestShuttle = 51,
    Forest5121 = 52,
    MotionBlindsLift = 65,
    MotionBlindsTiltLift = 66,
    MotionBlindsTilt = 67,
    CameraModule = 241,
    LocationModule = 242,
    SystemModule = 243,
    EnergyModule = 238,
    WeatherModule = 244,
    SecurityModule = 240,
}

export default Device;
export {DeviceType};